AppParams: relocate to application directory
