+++ xbmc/application/AppParams.cpp
