+++ xbmc/application/AppParams.h
