+++ xbmc/platform/darwin/ios/IOSEAGLView.mm
@@ -9,13 +9,13 @@
 #import "IOSEAGLView.h"
 
 #include "AppInboundProtocol.h"
-#include "AppParams.h"
 #include "Application.h"
 #import "IOSScreenManager.h"
 #include "ServiceBroker.h"
 #include "Util.h"
 #import "XBMCController.h"
 #include "application/AppEnvironment.h"
+#include "application/AppParams.h"
 #include "messaging/ApplicationMessenger.h"
 #include "settings/AdvancedSettings.h"
 #include "settings/SettingsComponent.h"
