+++ xbmc/platform/posix/utils/PosixInterfaceForCLog.cpp
@@ -8,8 +8,8 @@
 
 #include "PosixInterfaceForCLog.h"
 
-#include "AppParams.h"
 #include "ServiceBroker.h"
+#include "application/AppParams.h"
 
 #include <spdlog/sinks/dist_sink.h>
 #include <spdlog/sinks/stdout_color_sinks.h>
