+++ xbmc/platform/win10/Win10App.cpp
@@ -8,9 +8,9 @@
 
 #include "Win10App.h"
 
-#include "AppParams.h"
 #include "Application.h"
 #include "application/AppParamParser.h"
+#include "application/AppParams.h"
 #include "pch.h"
 #include "platform/Environment.h"
 #include "platform/xbmc.h"
